+++ src/runtime.ts
@@ -8,6 +8,8 @@ import { CompilerConfiguration } from './projects/compiler/compiler';
 import { PROJECTS } from './constants';
 import { GeneralCommands } from './commands';
 
+type ResetConifigurationMethod = () => Promise<void>;
+
 /**
  * Runtime class to manage workspace state and global variables.
  *
@@ -53,5 +55,22 @@ export abstract class Runtime {
   public static get compilerConfigurations(): CompilerConfiguration[] {
     const config = workspace.getConfiguration(PROJECTS.CONFIG.KEY);
     return config.get<CompilerConfiguration[]>(PROJECTS.CONFIG.COMPILER.CONFIGURATIONS, []);
+  }
+
+  public static async overrideConfiguration(section: string, key: string, value: any): Promise<ResetConifigurationMethod> {
+    const config = workspace.getConfiguration(section);
+    const previous = config.inspect(key);
+    if (!previous) return async () => {};
+    try {
+      if (previous.workspaceFolderLanguageValue !== undefined) return async () => await config.update(key, previous.workspaceFolderLanguageValue);
+      else if (previous.workspaceFolderValue !== undefined) return async () => await config.update(key, previous.workspaceFolderValue);
+      else if (previous.workspaceLanguageValue !== undefined) return async () => await config.update(key, previous.workspaceLanguageValue);
+      else if (previous.workspaceValue !== undefined) return async () => await config.update(key, previous.workspaceValue);
+      else if (previous.globalLanguageValue !== undefined) return async () => await config.update(key, previous.globalLanguageValue);
+      else if (previous.globalValue !== undefined) return async () => await config.update(key, previous.globalValue);
+      return async () => await config.update(key, undefined);
+    } finally {
+      await config.update(key, value);
+    }
   }
 }